MDL-49202 core_grades: perform same action when show/hiding in the form
If we hide on the main page, then visit the edit form and choose
to uncheck the 'hidden' checkbox and submit the form it will pass
'0' for the hidden parameter and 'false' for the cascade parameter to
the method 'set_hidden()' in the class 'grade_item'. If we pass
false and do not cascade, it doesn't go through all the grade_grades
for this item so the grades remains hidden to the students.